Finish testing

+++ note_keeper/views.py
@@ -1,52 +1,63 @@
 from django.http import HttpResponseRedirect
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import get_object_or_404, render, redirect
 from django.urls import reverse
 from django.views import generic
 from .models import Note
 from django.utils import timezone
 from django.contrib import messages
-from django.contrib.auth.mixins import LoginRequiredMixin
-
-# Create your views here.
+from django.http import Http404
 
 
-class IndexView(LoginRequiredMixin, generic.ListView):
-    login_url = '/accounts/login'
+class IndexView(generic.ListView):
     template_name = 'note_keeper/index.html'
     context_object_name = 'latest_note_list'
 
     def get_queryset(self):
         """Return the last five published Note."""
-        return Note.objects.order_by('-modify_date')[:3]
-      #  return Note.objects.filter(modify_date__lte=timezone.now()).order_by('-modify_date')[:]
+        return Note.objects.filter(modify_date__lte=timezone.now()).order_by('-modify_date')[:3]
 
     def get_context_data(self, **kwargs):
         context = super(IndexView, self).get_context_data(**kwargs)
         context['all_note'] = Note.objects.all()
         return context
 
-class DetailView(LoginRequiredMixin, generic.DetailView):
+class DetailView(generic.DetailView):
     model = Note
     template_name = 'note_keeper/note.html'
+
+    def get(self, request, *args, **kwargs):
+        try:
+            self.object = self.get_object()
+        except Http404:
+            messages.info(request, "There is no such note in our notebook")
+            return HttpResponseRedirect(reverse('note_keeper:index'))
+        context = self.get_context_data(object=self.object)
+        return self.render_to_response(context)
 
 class EditView(generic.DetailView):
     model = Note
     template_name = 'note_keeper/edit_note.html'
-    
+
 def delete_note(request, note_title):
-    messages.info(request, 'Note title #{} has been deleted'.format(note_title))
-    Note.objects.filter(pk=note_title).delete()
+    if Note.objects.filter(pk=note_title).exists(): 
+        messages.info(request, 'Note title #{} has been deleted'.format(note_title))
+        Note.objects.filter(pk=note_title).delete()
+        return HttpResponseRedirect(reverse('note_keeper:index'))
+    messages.info(request, 'Note title {} is not existed'.format(note_title))
     return HttpResponseRedirect(reverse('note_keeper:index'))
+
 
 def create_note(request):
     title = request.POST['title']
     content = request.POST['content']
-    for note in Note.objects.all():
-        if note.title == title:
-            messages.info(request, "Title is taken, please use different one!")
-            return HttpResponseRedirect(reverse('note_keeper:index'))
+    if not len(title) or not len(content):
+        messages.info(request, "Title or Content is empty, please write the form again")
+        return HttpResponseRedirect(reverse('note_keeper:index'))
 
-    note = Note(title=title, content=content)
+    if Note.objects.filter(pk=title).exists():
+        messages.info(request, 'Title is taken, remove the previous note first!')
+        return HttpResponseRedirect(reverse('note_keeper:index'))
+    note = Note.objects.create(title=title, content=content) 
     note.save()
     messages.info(request, 'Note: {} has been saved'.format(title))
     return HttpResponseRedirect(reverse('note_keeper:index'))
@@ -55,22 +66,33 @@ def update_note(request):
     title = request.POST['title']
     content = request.POST['content']
     curr_title = request.POST['old_title']
+    if len(title) < 3:
+        messages.info(request, "Please enter the title at least 3 characters or above")
+        return HttpResponseRedirect(reverse('note_keeper:index'))
+    if not Note.objects.filter(pk=curr_title).exists():
+        messages.info(request, "Sorry, I cannot let you do this")
+        return HttpResponseRedirect(reverse('note_keeper:index'))
     note = Note.objects.get(pk=curr_title)
+
+    if note.title == title and note.content == content:
+        messages.info(request, "No change detected")
+        # return redirect(request.META['HTTP_REFERER'])
+        return HttpResponseRedirect(reverse('note_keeper:index'))
+        
     if note.title != title: 
-        for db_note in Note.objects.all():
-            if db_note.title == title:
-                messages.info(request, "Title is taken, please use different one!")
-                return HttpResponseRedirect(reverse('note_keeper:note', args=[curr_title])) 
-        new_note = Note(title=title, content=content,doc=note.doc)  
+        if Note.objects.filter(pk=title):
+            messages.info(request, "Title is already taken! use different one.")
+            return redirect(request.META['HTTP_REFERER'])
+        new_note = Note(title=title, content=content,doc=note.doc, modify_date=timezone.now())  
         Note.objects.filter(pk=note.title).delete()
         new_note.save()
-    elif note.title == title and note.content == content:
-         messages.info(request, "No change detected")
-    else:
+        return HttpResponseRedirect(reverse('note_keeper:index'))
+    
+    if note.title == title and note.content != content:
         note.modify_date = timezone.now()
         note.title = title
         note.content = content 
         note.save()
-        messages.info(request, 'Note #{} has been updated'.format(title))
+    messages.info(request, 'Note #{} has been updated'.format(title))
     return HttpResponseRedirect(reverse('note_keeper:note', args=[title]))
